Added Socket class to avoid reuse of the code and ServerSession to manage each server connection separately
